+++ src/main/java/com/itplh/xhs/ui/User.java
@@ -1,4 +1,4 @@
-package com.itplh.xhs;
+package com.itplh.xhs.ui;
 
 import lombok.Data;
 
@@ -7,12 +7,12 @@ import lombok.Data;
  * @date: 2023-07-26 19:23
  */
 @Data
-public class AppUser {
+public class User {
 
     private String nickname;
     private String homeUrl;
 
-    public AppUser(String nickname, String homeUrl) {
+    public User(String nickname, String homeUrl) {
         this.nickname = nickname;
         this.homeUrl = homeUrl;
     }
